making c_fake to be of type int from long

+++ arunothia/fcgan.py
@@ -253,7 +253,7 @@ def main():
             mask = mask.type(torch.float)            
             noisy_label = torch.mul(1-mask, fake_label) + torch.mul(mask, real_label_smooth)
             
-            c_fake = torch.full(gen_class_labels.shape, c_fake_label, device=device, dtype=torch.long)
+            c_fake = torch.full(gen_class_labels.shape, c_fake_label, device=device, dtype=torch.int)
             d_fake_loss = (adversarial_loss(fake_pred, noisy_label) + auxiliary_loss(fake_aux, c_fake)) / 2
 
             # Total discriminator loss
